requirements.txt
Django~=4.2.0
django-allauth-2fa~=0.10.0
django-avatar~=7.1.1
django-allauth~=0.54.0
django-environ~=0.10.0
tldextract~=3.4.0
django-admin-rangefilter~=0.10.0
django-advanced-filters~=2.0.0
thefuzz~=0.19.0
python-Levenshtein~=0.20.9
docutils~=0.19
validators~=0.20.0
git+https://github.com/labcodes/django-el-pagination.git@django4.1
